backend/requirements.txt
django>=4.2,<5.0
channels>=4.0
channels-redis>=4.0
daphne>=4.0
websockets>=12.0
google-genai>=1.0
python-dotenv>=1.0
django-cors-headers>=4.0
